Prepare crc job for architecture repo
Create a job that can deploy the HCI VA with an extracted crc layout.
This change introduces the jobs definition and playbooks for such a job.
It creates and extracted crc layout and then calls networking mapper to
get the required networking config to deploy a VA. It needs to generate
and interfaces-info file, and translate the default network to ctlplane.

Additionally, it changes the 06-deploy-architecture playbook to work
with both controller and controller-0 nodes.

It makes use of scenario inhertance proposed in [1] to reuse most of the
existing HCI VA scenario.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/1491

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/1491