Create VMs with proper FQDN instead of shortname only
There is a need to get unique hostnames for the VM that will be mounting
some shared volumes in some infrastructure.

Allowing to get FQDN instead of shortnames allows to change the domain
of the VMs, making FQDN actually unique for each run.

The way we will leverage that feature is still unknown, but the support
will be in place whenever we need it.

A possibility would be to set the cifmw_devscripts_config.base_domain to
a unique value, or directly set the cifmw_reproducer_domain to whatever
value the user wants.